refactor: extract context logic into custom hooks
